fix: prometheus can return +/-Inf and NaN because it uses IEE754 floats
we can't return those via singer records because they don't have a proper
JSON encoding. Now it's possible to return them as string of course
but many targets have difficulty with that.

For the purposes of a simple tap, we simply map NaN and Inf to None/null instead, saving us this headache.